requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
sqlalchemy>=1.4,<1.5
databases==0.6.0
asyncpg==0.25.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==1.10.2
python-dotenv==1.0.1
psycopg2-binary==2.9.9
